refactor(migrate): extract modules from main.rs
Split 2090-line main.rs into four focused modules:
- template_diff: template variant diff/LCS computation
- cli: arg parsing and help text
- bib_postprocess: inferred bibliography postprocessing
- citation_validate: validate_and_normalize_inferred_citations

validate_and_normalize_inferred_citations split into
reject_invalid_inferred_citation and
normalize_inferred_citation_contributors, eliminating repeated
is_inferred_source checks and clarifying intent.

main.rs drops from 2090 to 545 lines of logic (tests remain inline).
